generate pkg/action
