+++ pkg/action/cl001/ac001/explainer.go
@@ -5,11 +5,13 @@ import (
 )
 
 func (e *Explainer) explain(ctx context.Context) (string, error) {
-	return `
+	s := `
 Check if the desired amount of Tenant Cluster master nodes are up and ready.
 
 	* Fetch all G8sControlPlane CRs spec.replicas so that we know how many masters the Tenant Cluster is supposed to have.
 	* Fetch the Tenant Cluster master nodes.
 	* Compare the current and desired amount of master nodes.
-	`, nil
+	`
+
+	return s, nil
 }
