+++ pkg/action/cl001/ac002/zz_generated.explainer.go
@@ -1,0 +1,28 @@
+package ac002
+
+import (
+	"context"
+
+	"github.com/giantswarm/microerror"
+)
+
+type ExplainerConfig struct {
+}
+
+type Explainer struct {
+}
+
+func NewExplainer(config ExplainerConfig) (*Explainer, error) {
+	e := &Explainer{}
+
+	return e, nil
+}
+
+func (e *Explainer) Explain(ctx context.Context) (string, error) {
+	s, err := e.explain(ctx)
+	if err != nil {
+		return "", microerror.Mask(err)
+	}
+
+	return s, nil
+}
